requirements.txt
streamlit
pillow
google-generativeai
dotenv
